# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /bot/reset-session:
    post:
      operationId: resetSession
      tags: [bot]
      summary: Reset session P&L counter
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SessionReset"
      responses:
        "200":
          description: Session reset
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  startingPnl:
                    type: number
                required:
                  - success
                  - startingPnl
...
  /positions/{dealId}:
    delete:
      operationId: closePosition
      tags: [positions]
      summary: Close a position
      parameters:
        - in: path
          name: dealId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Position closed
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  dealId:
                    type: string
                  profit:
                    type: number
                  message:
                    type: string
                required:
                  - success
                  - dealId
...
components:
  schemas:
...
    SessionReset:
      type: object
      properties:
        startingPnl:
          type: number
...